requirements.txt
mcp
openai
python-dotenv
vertexai
google-cloud-vision
google-genai
pdf2image
Pillow
ddgs
pypdf
requests
